[FDN-14] combine arrangements with order flow

+++ ConnectPlay.TicketPlay/Models/Order.cs
@@ -11,7 +11,7 @@ public record Order
     public ICollection<Ticket> Tickets { get; set; } = [];
     public ICollection<OrderArrangement> Arrangements { get; set; } = [];
     public OrderStatus Status { get; set; } = OrderStatus.Pending; // change to set beacuse we need to update the status and with init it cant be changed 
-    public required float Total { get; set; }
+    public required decimal Total { get; set; }
     public string OrderCode { get; set; } = GenerateOrderCode();
 
     // generate something that resembles a unique code based on the GUID of the first ticket in the order
